+++ serve/routes/routes.product.js
@@ -26,26 +26,30 @@ routesProducts.get("/", async (req, res) => {
   var pageSize = parseInt(req.query.pageSize) || 20;
   const query = queryParams(req.query);
   try {
-    await ModelProduct.find(query)
-      .sort({ update_at: -1 })
-      .skip(pageIndex * pageSize) //Notice here
-      .limit(pageSize)
-      .exec((err, doc) => {
-        if (err) {
-          return res.json(err);
-        }
-        ModelProduct.countDocuments(query).exec((count_error, count) => {
+    if (req.headers.authorization) {
+      await ModelProduct.find(query)
+        .sort({ update_at: -1 })
+        .skip(pageIndex * pageSize) //Notice here
+        .limit(pageSize)
+        .exec((err, doc) => {
           if (err) {
-            return res.json(count_error);
+            return res.json(err);
           }
-          return res.json({
-            total: count,
-            pageIndex,
-            pageSize,
-            data: doc,
+          ModelProduct.countDocuments(query).exec((count_error, count) => {
+            if (err) {
+              return res.json(count_error);
+            }
+            return res.json({
+              total: count,
+              pageIndex,
+              pageSize,
+              data: doc,
+            });
           });
         });
-      });
+    } else {
+      res.status(401).json({ message: "Không có quyền", statusCode: 401 });
+    }
   } catch (error) {
     res.status(500).json({ statusCode: 500, message: error.message });
   }
@@ -53,16 +57,22 @@ routesProducts.get("/", async (req, res) => {
 
 routesProducts.get("/:id", async (req, res) => {
   try {
-    // if (req.headers.authorization) {
-    const id = req.params.id;
-    if (id) {
-      const data = await ModelProduct.findById(id);
+    if (req.headers.authorization) {
+      const id = req.params.id;
+      if (id) {
+        const data = await ModelProduct.findById(id);
 
-      if (data) {
-        res.status(200).json({
-          statusCode: 200,
-          data,
-        });
+        if (data) {
+          res.status(200).json({
+            statusCode: 200,
+            data,
+          });
+        } else {
+          res.status(400).json({
+            statusCode: 400,
+            message: "Không tìm thấy product",
+          });
+        }
       } else {
         res.status(400).json({
           statusCode: 400,
@@ -70,10 +80,7 @@ routesProducts.get("/:id", async (req, res) => {
         });
       }
     } else {
-      res.status(400).json({
-        statusCode: 400,
-        message: "Không tìm thấy product",
-      });
+      res.status(401).json({ message: "Không có quyền", statusCode: 401 });
     }
   } catch (error) {
     res.status(500).json({ message: error.message + req.params });
@@ -116,33 +123,36 @@ routesProducts.post("/", async (req, res) => {
     resolutionFrontCamera: req.body.resolutionFrontCamera ?? "",
   });
   try {
-    // if (req.headers.authorization) {
-    const requiredValue = Object.keys(req.body).reduce((obj, item) => {
-      if (
-        item === "name" ||
-        item === "price" ||
-        item === "quantity" ||
-        item === "image" ||
-        item === "categoryId"
-      ) {
-        if (req.body[item] === "") {
-          obj[item] = `${item} không được bỏ trống`;
+    if (req.headers.authorization) {
+      const requiredValue = Object.keys(req.body).reduce((obj, item) => {
+        if (
+          item === "name" ||
+          item === "price" ||
+          item === "quantity" ||
+          item === "image" ||
+          item === "categoryId"
+        ) {
+          if (req.body[item] === "") {
+            obj[item] = `${item} không được bỏ trống`;
+          }
+          return obj;
         }
-        return obj;
-      }
-    }, {});
+      }, {});
 
-    if (Object.keys(requiredValue).length === 0) {
-      await data.save();
-      res.status(200).json({
-        statusCode: 200,
-        message: "Tạo mới product thành công",
-      });
+      if (Object.keys(requiredValue).length === 0) {
+        await data.save();
+        res.status(200).json({
+          statusCode: 200,
+          message: "Tạo mới product thành công",
+        });
+      } else {
+        res.status(200).json({
+          statusCode: 422,
+          data: requiredValue,
+        });
+      }
     } else {
-      res.status(200).json({
-        statusCode: 422,
-        data: requiredValue,
-      });
+      res.status(401).json({ message: "Không có quyền", statusCode: 401 });
     }
   } catch (error) {
     res.status(500).json({ statusCode: 500, message: error.message });
@@ -236,15 +246,19 @@ routesProducts.put("/:id", async (req, res) => {
 //Delete by ID Method
 routesProducts.delete("/:id", async (req, res) => {
   try {
-    const id = req.params.id;
-    const data = await ModelProduct.findByIdAndDelete(id);
-    if (data) {
-      res.send({
-        message: "Xóa product thành công",
-        statusCode: 200,
-      });
+    if (req.headers.authorization) {
+      const id = req.params.id;
+      const data = await ModelProduct.findByIdAndDelete(id);
+      if (data) {
+        res.send({
+          message: "Xóa product thành công",
+          statusCode: 200,
+        });
+      } else {
+        res.status(400).json({ statusCode: 400, message: error.message });
+      }
     } else {
-      res.status(400).json({ statusCode: 400, message: error.message });
+      res.status(401).json({ message: "Không có quyền", statusCode: 401 });
     }
   } catch (error) {
     res.status(500).json({ statusCode: 500, message: error.message });
